feat: Add Support for a HTTP Callback.
On response, it handles the HTTP status as follows:

HTTP 200-299: Client is allowed to play the path, continue as normal and run through the entire rest of the setup procedure without any changes
HTTP 300-399: Client is allowed to play the path, but from a different configured source: than the one defined in the configuration. The HTTP backend should include a location HTTP header with the new proxy source, if that's not the case, close the rtsp client connection.
HTTP 400-499: Client is not allowed to play the path, reject playback.
Any other HTTP code or no HTTP response: Close the connection